fix(lingbot): fp32 VAE decode + rich default negative prompt
Two correctness fixes found by comparing served output against the upstream
LingBotVideoPipeline (the port was faithful, but two usage defaults diverged
from the reference and degraded quality):

- VAE decode now runs in fp32 (weights + activations), matching the upstream
  runner's --vae_dtype fp32. It was built bf16 and decoded under a bf16
  autocast, which visibly degrades frames.
- default_negative_prompt is now the upstream structured-JSON negative (verbatim
  in default_negative_prompt.py) instead of ""; an empty negative gives weak
  CFG. Applied when a request omits negative_prompt.

Also documents on the adapter that LingBot expects structured-JSON prompts
(rewriter output), not raw text — raw prompts garble on the reference model too;
serving the 27B rewriter is out of scope for this port.

Verified on one GPU: served path renders a coherent 81-frame 832x480 clip from a
structured prompt, matching the upstream baseline.

+++ mstar/model/lingbot/weight_loader.py
@@ -166,7 +166,9 @@ def build_lingbot_vae_decoder(
     with torch.device("meta"):
         vae = LingBotWanVaeDecoder(config)
     vae.to_empty(device=device)
-    vae.to(dtype=torch.bfloat16)
+    # The Wan 3D-VAE decode is run in fp32 (upstream inference uses --vae_dtype fp32);
+    # bf16 here visibly degrades decoded frames, so keep VAE weights in fp32.
+    vae.to(dtype=torch.float32)
     _load_exact(
         vae,
         _iter_diffusers_safetensors(vae_dir, device),
